+++ .github/CHANGELOG.md
@@ -10,6 +10,7 @@ This update brings some services, features and customizations to the project. Se
 - Added `twMerge()` (tailwind merge) dev dependency to help working with dynamic classes.
 - Added `CryptoHash` static class for handling encryptation related tasks.
 - Added a custom react hook `useViewport()` to handle real-time width and height of the current viewport.
+- Added `DataValidation` static class for handling data validation tasks.
 
 ### UI and Components
 - Added the Theme context to handle light and dark mode on the application.